fix: resolve all clippy -D warnings
- mock.rs: .as_ref().map(|v| v.clone()) → .clone() (useless_asref)
- observe.rs: manual if/div → checked_div().unwrap_or(0) (manual_checked_ops)
- restbridge.rs: 200|201|202 → 200..=202 (manual_range_patterns)
- udsbr.rs: (sid | 0x40) as u8 → sid | 0x40 (unnecessary_cast, sid is already u8)

+++ src/udsbr.rs
@@ -73,7 +73,7 @@ impl Controller for UdsBridge {
         let sid = cmd_to_uds_sid(cmd.cmd_type);
         let data = cmd.payload.as_deref().unwrap_or(&[]);
         let resp_data = self.transport.request(sid, data, timeout)?;
-        let pos_resp = (sid | 0x40) as u8; // UDS positive response SID
+        let pos_resp = sid | 0x40; // UDS positive response SID
         let status = if resp_data.first() == Some(&pos_resp) {
             ResponseStatus::OK
         } else {
